requirements.txt
ollama
fastapi
uvicorn
logging
python-docx 
chardet 
beautifulsoup4 
extract_msg